+++ integrated_channels/sap_success_factors/exporters/learner_data.py
@@ -73,20 +73,6 @@ class SapSuccessFactorsLearnerExporter(LearnerExporter):
                     enterprise_customer_uuid=self.enterprise_configuration.enterprise_customer.uuid,
                     plugin_configuration_id=self.enterprise_configuration.id
                 ),
-                SapSuccessFactorsLearnerDataTransmissionAudit(
-                    enterprise_course_enrollment_id=enterprise_enrollment.id,
-                    sapsf_user_id=sapsf_user_id,
-                    user_email=enterprise_enrollment.enterprise_customer_user.user_email,
-                    course_id=enterprise_enrollment.course_id,
-                    course_completed=course_completed,
-                    completed_timestamp=completed_date,
-                    sap_completed_timestamp=sap_completed_timestamp,
-                    grade=grade,
-                    total_hours=total_hours,
-                    credit_hours=total_hours,
-                    enterprise_customer_uuid=self.enterprise_configuration.enterprise_customer.uuid,
-                    plugin_configuration_id=self.enterprise_configuration.id
-                ),
             ]
         LOGGER.info(
             generate_formatted_log(