+++ tests/test_integrated_channels/test_degreed/test_exporters/test_learner_data.py
@@ -81,9 +81,8 @@ class TestDegreedLearnerExporter(unittest.TestCase):
             completed_date=completed_date,
             course_completed=course_completed,
         )
-        assert len(learner_data_records) == 2
+        assert len(learner_data_records) == 1
         assert learner_data_records[0].course_id == self.course_key
-        assert learner_data_records[1].course_id == self.course_id
 
         for learner_data_record in learner_data_records:
             assert learner_data_record.enterprise_course_enrollment_id == enterprise_course_enrollment.id