+++ CHANGELOG.rst
@@ -17,6 +17,10 @@ Unreleased
 ----------
 Nothing unreleased.
 
+[4.7.2]
+--------
+feat: stop learner data transmissions for course runs
+
 [4.7.1]
 --------
 chore: retire Degreed v1 code from the set of channels